refactor: 이제 사용되지 않는 request를 제거하고 request_temp를 request로 리네임
