+++ utils/api/LibraryBookStockApi.ts
@@ -1,7 +1,7 @@
 import { LoanInfo } from "@/types/Loan";
 import { StockLoanStateApiSpec } from "@/types/ApiSpec";
 import { convertLoanInfo } from "./ApiResponseConvertor";
-import { getApiClient, Side } from "./common/Request_TEMP";
+import { getApiClient, Side } from "./common/Request";
 
 interface Props {
   libraryId: string;
